% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{conditional_SMC}
\alias{conditional_SMC}
\title{Conditional SMC algorithm}
\usage{
conditional_SMC(n_particles, observations, cond_trajectory, emission_density,
  transition_sam, initial_sam, ...)
}
\arguments{
\item{n_particles}{number of particles}

\item{observations}{a vector of observations on which to perform inference}

\item{cond_trajectory}{a vector of states representing the trajectory on which to condition}

\item{emission_density}{a function of the form f(y,x) giving the emission density of y|x}

\item{transition_sam}{a function of the form f(x) sampling a new position, i.e. propagating the particle}

\item{initial_sam}{a function of the form f() sampling the initial positions}

\item{...}{arguments to be passed to emission_density, transition_sam and initial_sam
(e.g. for Ornstein-Uhlenbeck process, provide values for sigma and delta)}
}
\value{
a list containing matrices of positions and weights
(rows represent generations/time steps, columns represent individuals/particles)
and an ancestry matrix (where element (i,j) is the parent in generation i of individual j in generation i+1)
}
\description{
Conditional SMC implementation with multinomial resampling for a generic (exchangeable) HMM
}
\author{
Suzie Brown
}
