% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_sim.R
\name{moranSim}
\alias{moranSim}
\title{Moran Sample}
\usage{
moranSim(N, N.gen, rescale = FALSE)
}
\arguments{
\item{N}{constant population size}

\item{N.gen}{number of generations (single reproductions) to simulate}

\item{rescale}{return only every Nth generation? (for comparability with W-F model)}
}
\value{
an object of class 'genealogy'
}
\description{
Generate a realisation of the Moran population model
}
\details{
The attribute 'history' contains an (N.gen-1)xN matrix , where the (i,j)th entry specifies the parent of individual j of generation i+1. For comparability to Wright-Fisher model, multiply N.gen by N: one generation in W-F model comprises N reproduction events.
}
\author{
Suzie Brown
}
