% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OU_RTS_smooth.R
\name{ou_rts_smooth}
\alias{ou_rts_smooth}
\title{RTS smoother for OU model}
\usage{
ou_rts_smooth(observations, delta, sigma, prior_mean = 0, prior_var = 1)
}
\arguments{
\item{observations}{vector of observations from the OU model}

\item{delta}{drift parameter of the Ornstein-Uhlenbeck process}

\item{sigma}{noise parameter of the Ornstein-Uhlenbeck process}

\item{prior_mean}{prior mean}

\item{prior_var}{prior variance}
}
\value{
list containing the smoothed means and variances
}
\description{
Calculate the Rauch-Tung-Striebel smoothing distribution for observations from an Ornstein-Uhlenbeck model
}
\author{
Suzie Brown
}
