% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OU_sim.R
\name{ou_sim}
\alias{ou_sim}
\title{Ornstein-Uhlenbeck simulation}
\usage{
ou_sim(n_obs, delta, sigma, return_states = FALSE)
}
\arguments{
\item{n_obs}{the number of observations (excluding initial state) to produce, i.e. the number of time steps}

\item{delta}{drift parameter of the Ornstein-Uhlenbeck process}

\item{sigma}{noise parameter of the Ornstein-Uhlenbeck process}

\item{return_states}{return the hidden states as well as the noisy observations?}
}
\value{
a vector of (n_obs + 1) observations; or if return_states = TRUE, a list containing a vector of hidden
states and a corresponding vector of observations
}
\description{
Simulate synthetic data from the Ornstien-Uhlenbeck model
}
\author{
Suzie Brown
}
