% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy_plot.R
\name{plot.genealogy}
\alias{plot.genealogy}
\title{Genealogy Plot}
\usage{
\method{plot}{genealogy}(history = attr(history, "history"),
  title = attr(history, "model"), col.line = "darkgrey",
  shade.dead = !highlight.sample, col.alive = "black", col.dead = "grey",
  highlight.sample = FALSE, sampl = 1:N, col.highlight = "purple", ...)
}
\arguments{
\item{history}{an object of class 'genealogy' as returned from WFsim or moranSim, where the (i,j)th entry specifies the parent of individual j of generation i+1}

\item{title}{title for the plot}

\item{col.line}{the colour used to plot parent-offspring lines}

\item{shade.dead}{use a different colour to shade individuals with no offspring (dead lineages)?}

\item{col.alive}{if shade.dead=TRUE, the colour to shade individuals with offspring}

\item{col.dead}{if shade.dead=TRUE, the colour to shade individuals without offspring}

\item{highlight.sample}{highlight the ancestry of a sample from the most recent generation?}

\item{sampl}{if highlight.sample=TRUE, which individuals from the most recent generation to include in the sample}

\item{col.highlight}{if highlight.sample=TRUE, the colour used to plot the highlighted ancestry}
}
\description{
Plot method for objects of class 'genealogy'
}
\author{
Suzie Brown
}
