% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy_plot.R
\name{plot.samplegenealogy}
\alias{plot.samplegenealogy}
\title{Sampled Genealogy Plot}
\usage{
\method{plot}{samplegenealogy}(samplehistory)
}
\arguments{
\item{samplehistory}{an object of class samplegenealogy}
}
\description{
Plot method for objects of class samplegenealogy
}
\author{
Suzie Brown
}
