% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_ancestry.R
\name{traceAncestry}
\alias{traceAncestry}
\title{Trace sample ancestry}
\usage{
traceAncestry(history, sampl)
}
\arguments{
\item{history}{an ocject of class genealogy, as returned from WFsim or moranSim}

\item{sampl}{a vector of integers indicating which individuals from the most recent generation are to be included in the sample}
}
\value{
an object of class samplegenealogy
}
\description{
Trace the ancestry of individuals in a sample and find their most recent common ancestor
}
\author{
Suzie Brown
}
