% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_sim.R
\name{WFplot}
\alias{WFplot}
\title{Wright-Fisher Plot}
\usage{
WFplot(history, col.line = "darkgrey", shade.dead = TRUE,
  col.alive = "black", col.dead = "grey", highlight.sample = FALSE,
  sampl = 1:N, col.highlight = "purple")
}
\arguments{
\item{history}{a matrix as returned from WFsim, where the (i,j)th entry specifies the parent of individual j of generation i+1}

\item{col.line}{the colour used to plot parent-offspring lines}

\item{shade.dead}{use a different colour to shade individuals with no offspring (dead lineages)?}

\item{col.alive}{if shade.dead=TRUE, the colour to shade individuals with offspring}

\item{col.dead}{if shade.dead=TRUE, the colour to shade individuals without offspring}

\item{highlight.sample}{highlight the ancestry of a sample from the most recent generation?}

\item{sampl}{if highlight.sample=TRUE, which individuals from the most recent generation to include in the sample}

\item{col.highlight}{if highlight.sample=TRUE, the colour used to plot the highlighted ancestry}
}
\description{
Plot a realisation of the Wright-Fisher population model
}
\author{
Suzie Brown
}
