% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_sim.R
\name{WFsim}
\alias{WFsim}
\title{Wright-Fisher Sample}
\usage{
WFsim(N, N.gen)
}
\arguments{
\item{N}{constant population size}

\item{N.gen}{number of generations to simulate}
}
\value{
an (N.gen-1)xN matrix , where the (i,j)th entry specifies the parent of individual j of generation i+1
}
\description{
Generate a realisation of the Wright-Fisher population model
}
\author{
Suzie Brown
}
